Reduce use of very long lines in .mdx (#134)
These long lines are hard to read in a wide window
and make it hard to read diffs, because a single
change in a long line causes the whole line to be
marked different.

This change has no content changes, only formatting.

There is no requirement going forward to keep lines
equal in length when editing.  This would also cause
diffs to be larger than required.